Move fixtures file
