+++ src/app/routes/article/fixtures.js
